+++ src/matterbridgePlatform.ts
@@ -27,6 +27,7 @@ import path from 'node:path';
 
 // Matter
 import { EndpointNumber } from '@matter/main';
+import { Descriptor } from '@matter/main/clusters/descriptor';
 // Node AnsiLogger module
 import { AnsiLogger, CYAN, db, er, LogLevel, nf, wr } from 'node-ansi-logger';
 // Node Storage module
@@ -36,6 +37,7 @@ import { NodeStorage, NodeStorageManager } from 'node-persist-manager';
 import { Matterbridge } from './matterbridge.js';
 import { MatterbridgeEndpoint } from './matterbridgeEndpoint.js';
 import { checkNotLatinCharacters } from './matterbridgeEndpointHelpers.js';
+import { bridgedNode } from './matterbridgeDeviceTypes.js';
 import { isValidArray, isValidObject, isValidString } from './utils/export.js';
 
 // Platform types
@@ -290,6 +292,14 @@ export class MatterbridgePlatform {
       );
       return;
     }
+    if (!device.deviceName) {
+      this.log.error(`Device with uniqueId ${CYAN}${device.uniqueId}${er} has no deviceName. The device will not be added.`);
+      return;
+    }
+    if (!device.serialNumber) {
+      this.log.error(`Device with uniqueId ${CYAN}${device.uniqueId}${er} has no serialNumber. The device will not be added.`);
+      return;
+    }
     if (device.deviceName && this._registeredEndpointsByName.has(device.deviceName)) {
       this.log.error(`Device with name ${CYAN}${device.deviceName}${er} is already registered. The device will not be added. Please change the device name.`);
       return;
@@ -298,20 +308,37 @@ export class MatterbridgePlatform {
       this.log.debug(`Device with name ${CYAN}${device.deviceName}${db} has non latin characters.`);
     }
 
-    /* TODO: Do we need to handle the childbridged mode with DynamicPlatform here if the implementation forgets to call createDefaultBridgedDeviceBasicInformationClusterServer?
-    if (this.matterbridge.bridgeMode === 'childbridge' && this.type === 'DynamicPlatform' && device.mode === undefined) {
+    /**
+     * We check if the implementation called createDefaultBasicInformationClusterServer() instead of createDefaultBridgedDeviceBasicInformationClusterServer()
+     * This is correct with Accessory platforms so we check if we are running in bridge mode and add the bridgedNode.
+     */
+    if (device.mode === undefined && (this.matterbridge.bridgeMode === 'bridge' || (this.matterbridge.bridgeMode === 'childbridge' && this.type === 'DynamicPlatform')) && !device.deviceTypes.has(bridgedNode.code)) {
+      // If the device is a bridged device, we add the bridgedNode to the deviceTypes map
+      device.deviceTypes.set(bridgedNode.code, bridgedNode);
+      // We add also the bridgedNode to the Descriptor Cluster options
       const options = device.getClusterServerOptions(Descriptor.Cluster.id);
       if (options) {
         const deviceTypeList = options.deviceTypeList as { deviceType: number; revision: number }[];
-        deviceTypeList.push({ deviceType: bridgedNode.code, revision: bridgedNode.revision });
+        if (!deviceTypeList.find((dt) => dt.deviceType === bridgedNode.code)) {
+          deviceTypeList.push({ deviceType: bridgedNode.code, revision: bridgedNode.revision });
+        }
       }
-      device.createDefaultBridgedDeviceBasicInformationClusterServer(deviceName, serialNumber, vendorId, vendorName, productName, softwareVersion, softwareVersionString, hardwareVersion, hardwareVersionString);
+      device.createDefaultBridgedDeviceBasicInformationClusterServer(
+        device.deviceName,
+        device.serialNumber,
+        device.vendorId,
+        device.vendorName,
+        device.productName,
+        device.softwareVersion,
+        device.softwareVersionString,
+        device.hardwareVersion,
+        device.hardwareVersionString,
+      );
     }
-    */
 
     await this.matterbridge.addBridgedEndpoint(this.name, device);
-    if (device.uniqueId) this._registeredEndpoints.set(device.uniqueId, device);
-    if (device.deviceName) this._registeredEndpointsByName.set(device.deviceName, device);
+    this._registeredEndpoints.set(device.uniqueId, device);
+    this._registeredEndpointsByName.set(device.deviceName, device);
   }
 
   /**